Adds basic online gameplay
- Adds socket.io and uuidv4 as dependencies.

- Adds a room selection page and css needed to style it.

- Allows for selection between online, local multiplayer, and
singleplayer game modes. Singleplayer allows for 3 AI difficulties.

- Separates script for instantiating socket.io and initializing the game
so as to allow for the aforementioned multiple game types.

- Adds server side logic for handling several socket.io events for
management both on the socket.io server and on the local server.

- To implement this, adds two classes Room and RoomManager to the server
models, allowing for in-memory management of game rooms. Integrates the
logic with socket.io rooms.

- Implements ejs templating and partials to allow for server-side page
rendering. Implements locals to pass room/game data to the client
scripts.

- Adds routes with API endpoints necessary for serving the pages and
relevant scripts.

- Adds many comments throughout application.