Adds API tests and makes several changes to app
- Adds ChaiHttp as a dependency and sets up testing script in
package.json.

- Adds several API tests under /tests/server/routes/index.test.js.

- Adds mocha to eslintrc.json env.

- Removes the setup of roomId/name variables by passing locals to an EJS
template. The variables were set up to allow socket.io clients to know
which room to join for an online game. Replaces this by redirecting
successful room joins to a /room/:roomId API that allows for a
client-side document.URL query and using the :roomId parameter to create
or join a socket.io room.

- Removes the restart button from the index.ejs template and instead
adds it via DOM manipulation when singleplayer games are instantiated.

- Removes unnecessary "createRoom" socket.io event from socketSetup.js.
Replaces this with an API which will be used for DB manipulation when a
DB is set up. Consequently, adds logic to "joinRoom" event to handle
room creation.

- Removes several locals variable declarations throughout
/routes/index.js. Adds several endpoints related to room management.
Will use these to integrate with a database.

- Adds functions to Room and RoomManager classes to implement basic CRUD
functions from API interaction.

- Changes Room constructor to only require a name, and removes a players
requirement from RoomManager.addRoom().

- Exports a server instance and a RoomManager instance from start.js for
testing purposes.

- Updates singleplayer.js and joinRoom.js to get the difficulty/roomId,
respectively, from the current page's URL and use them to instantiate
their respective game type, and, in the case of joinRoom.js, create/join
a socket.io room.

- Changes selection.io.js to fetch rooms from the server directly and
not through socket.io. Also adds a fontawesome lock icon to signal that
a room is password protected.

- Reconfigures room creation to occur on the index page and not on a
seperate page.

- As such moves the room creation modal and styles to the "/" page.

- Adds a helper script consisting of a single function that gets the
last part of the url. Used to set difficulty and create socket.io rooms
via the processes described above.